REMAKE II: login ok. Fazer tarefas

+++ projetoFazendo/projetoCasamentoWeb/src/main/java/br/com/ifpe/tads/projetoCasamentoWeb/repository/Repository.java
@@ -48,7 +48,7 @@ public abstract class Repository<Entidade> {
 		em.remove(em.merge(entidade));
 	}
 
-	public Entidade buscar(Integer idEntidade) {
+	public Entidade buscar(Long idEntidade) {
 		return em.find(classe, idEntidade);
 	}
